Gate on the wasm family instead of the wasm32 arch (#39)
* Gate on the wasm family instead of the wasm32 arch

Every target gate spelled out `target_arch = wasm32`, so on
`wasm64-unknown-unknown` the alias silently fell back to native `tokio`
and the glue never compiled. `target_family = wasm` covers both
widths, while the unchanged vendor and OS checks keep Emscripten and
WASI out.

The task pointer that crosses the worker boundary is now sent as
`usize` instead of `u32`, so it isn't truncated on a 64-bit wasm
target. On `wasm32` this compiles to the exact same thing.

Note that actually running on `wasm64-unknown-unknown` still waits on
`wasm-bindgen` support for the target; this only stops the crate from
picking the wrong side of the alias there.

Closes #26.

* Require wasm-bindgen 0.2.120 for the wasm64 pointer ABI

* Correct the wasm64 support note in the READMEs

* Shorten the wasm64 note in the READMEs

* Add a wasm64 clippy job to CI

* Rename the web CI job to web32

+++ package/src/glue/task/pool.rs
@@ -216,7 +216,9 @@ impl WorkerPool {
     let worker = self.get_worker()?;
     let work = Box::new(task);
     let ptr = Box::into_raw(work);
-    match worker.post_message(&JsValue::from(ptr as u32)) {
+    // `usize`, not `u32`, so that the pointer survives on `wasm64`.
+    // It crosses JS as an `f64` there, which is exact below 2^53.
+    match worker.post_message(&JsValue::from(ptr as usize)) {
       Ok(()) => Ok(worker),
       Err(error) => {
         unsafe {
@@ -424,7 +426,7 @@ impl PoolState {
 
 /// Entry point invoked by JavaScript in a worker.
 #[wasm_bindgen]
-pub fn task_worker_entry_point(ptr: u32) -> Result<(), JsValue> {
+pub fn task_worker_entry_point(ptr: usize) -> Result<(), JsValue> {
   let ptr = unsafe { Box::from_raw(ptr as *mut Task) };
   let global = global().unchecked_into::<DedicatedWorkerGlobalScope>();
   (ptr.callable)();
